[cxx-interop] Workaround a template instantiation issue, part 2
This is a follow-up to d483a8ab.

rdar://145512778
